+++ src/app/cart/components/cart-item/cart-item.component.ts
@@ -7,8 +7,9 @@ import {ProductInCart} from '../../../models/ProductInCartModel';
   templateUrl: './cart-item.component.html',
   styleUrls: ['./cart-item.component.css']
 })
-export class CartItemComponent implements OnInit, DoCheck {
+export class CartItemComponent implements OnInit {
   @Input() productInCart: ProductInCart;
+
   @Output() incrementQuantity: EventEmitter<void> = new EventEmitter<void>();
   @Output() decrementQuantity: EventEmitter<void> = new EventEmitter<void>();
   @Output() delete: EventEmitter<void> = new EventEmitter<void>();
@@ -19,13 +20,6 @@ export class CartItemComponent implements OnInit, DoCheck {
 
   ngOnInit() {
     this.productInCartDiffer = this.keValueDiffers.find(this.productInCart).create();
-  }
-
-  ngDoCheck() {
-    const changes = this.productInCartDiffer.diff(this.productInCart);
-    if (changes) {
-      changes.forEachChangedItem(change => this.productInCart[change.key] = change.currentValue);
-    }
   }
 
   onIncrementQuantity() {