+++ src/app/core/core.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {ConstantsService} from './services/constants.service';
+import {GeneratorService} from './services/generator.service';
+import {LocalStorageService} from './services/local-storage.service';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [],
+  providers: [{
+    provide: ConstantsService.token,
+    useValue: ConstantsService.value
+  }, {
+    provide: GeneratorService.token,
+    useFactory: GeneratorService.factory(5)
+  },
+  LocalStorageService
+  ]
+})
+export class CoreModule { }